+++ NEWS.md
@@ -1,4 +1,4 @@
-# nominatimlite (development version)
+# nominatimlite 0.3.0
 
 -   Add a `progressbar` parameter to `geo_lite()`, `geo_lite_sf()`,
     `reverse_geo_lite()` and `reverse_geo_lite_sf()` to display progress in the